requirements.txt
asgiref==3.8.1
Django==5.1.4
pillow==11.0.0
qrcode==8.0
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
